// src/app/modules/sidenav/components/draw/draw-modules/draw-options/enum/draw-options.enum.ts
export enum DrawTools {
	Point = "drawPoint",
	Line = "drawLine",
	Polygon = "drawPolygon",
}

export enum ProjectionType {
	Metric = "metric",
	Degree = "degree",
}

export enum DrawShapes {
	Circle = "Circle",
	Square = "Square",
	Rectangle = "Rectangle",
	Triangle = "Triangle",
	Arrow = "Arrow",
}